fusion solvers.cs
