--- Sudoku.HumanTechnique/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Sudoku.HumanTechnique
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
